+++ lib/quick_js_ex/native.ex
@@ -1,6 +1,25 @@
 defmodule QuickJSEx.Native do
   @moduledoc false
-  use Rustler, otp_app: :quickjs_ex, crate: "quickjs_ex_nif"
+
+  version = Mix.Project.config()[:version]
+
+  use RustlerPrecompiled,
+    otp_app: :quickjs_ex,
+    crate: "quickjs_ex_nif",
+    base_url: "https://github.com/dannote/quickjs_ex/releases/download/v#{version}",
+    force_build: System.get_env("QUICKJS_EX_BUILD") in ["1", "true"],
+    targets: ~w(
+      aarch64-apple-darwin
+      aarch64-unknown-linux-gnu
+      aarch64-unknown-linux-musl
+      x86_64-apple-darwin
+      x86_64-unknown-linux-gnu
+      x86_64-unknown-linux-musl
+      x86_64-pc-windows-gnu
+      x86_64-pc-windows-msvc
+    ),
+    nif_versions: ["2.15"],
+    version: version
 
   @spec start_runtime(pid(), boolean()) :: :ok
   def start_runtime(_pid, _browser_stubs), do: :erlang.nif_error(:nif_not_loaded)